requirements.txt
agno==1.3.4
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.16
duckduckgo_search==8.0.1
fastapi==0.115.12
frozendict==2.4.6
gitdb==4.0.12
GitPython==3.1.44
groq==0.22.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
jiter==0.9.0
lancedb==0.21.2
lxml==5.3.2
markdown-it-py==3.0.0
mdurl==0.1.2
multitasking==0.0.11
numpy==2.2.5
openai==1.75.0
overrides==7.7.0
packaging==25.0
pandas==2.2.3
peewee==3.17.9
phidata==2.7.10
platformdirs==4.3.7
primp==0.15.0
pyarrow==19.0.1
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
Pygments==2.19.1
pylance==0.25.2
pypdf==5.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
tantivy==0.22.2
tomli==2.2.1
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
yfinance==0.2.55
